+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/BanyanDBStorageConfig.java
@@ -106,6 +106,7 @@ public class BanyanDBStorageConfig extends ModuleConfig {
         private int segmentQueryMaxSize = 200;
         private int profileDataQueryBatchSize = 100;
         private boolean cleanupUnusedTopNRules = true;
+        private String namespace = "sw";
     }
 
     // The configuration of the groups.